+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file[^1].
 
 ### Updated
 - migrate factories to laravel 8 + style factories
+- article thumbnail according to the new design
 
 ## [2.65.0] - 2022-09-12
 ### Fixed